+++ blueocean-git-pipeline/src/main/java/io/jenkins/blueocean/blueocean_git_pipeline/GitPipelineCreateRequest.java
@@ -4,8 +4,8 @@ import com.cloudbees.plugins.credentials.common.StandardCredentials;
 import com.google.common.collect.Lists;
 import io.jenkins.blueocean.commons.ErrorMessage;
 import io.jenkins.blueocean.commons.ErrorMessage.Error;
-import io.jenkins.blueocean.pipeline.api.AbstractMultiBranchCreateRequest;
 import io.jenkins.blueocean.rest.model.BlueScmConfig;
+import io.jenkins.blueocean.scm.api.AbstractMultiBranchCreateRequest;
 import jenkins.branch.MultiBranchProject;
 import jenkins.model.Jenkins;
 import jenkins.plugins.git.GitSCMSource;
