+++ blueocean-pipeline-scm-api/src/main/java/io/jenkins/blueocean/scm/api/AbstractPipelineCreateRequest.java
@@ -1,4 +1,4 @@
-package io.jenkins.blueocean.pipeline.api;
+package io.jenkins.blueocean.scm.api;
 
 import hudson.model.Item;
 import hudson.model.Items;
